minor warning fix to app_ffview
git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@2041 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/app_ffview/Viewer.cpp
@@ -201,7 +201,7 @@ bool Viewer::setParam(string param, double value)
   if(param == "set_base_ipf")
     m_base_ipf = value;
   else if(param == "uniform_amount")
-    m_piece_count = value;
+    m_piece_count = (int)(value);
   else if(param == "mod_base_ipf")
     m_base_ipf += value;
   else if(param == "set_base_aof")
